Remove unused interfaces
BuildDao, RemoteLogin, SummaryDao, WebLinks
