+++ src/dal/builddao.ts
@@ -1,7 +1,25 @@
 "use strict";
 
+import {WebLinks} from "./webLinks";
 import {Build} from "../bll/entities/build";
+import {TYPES} from "../bll/utils/constants";
+import {XmlParser} from "../bll/utils/xmlparser";
+import {inject, injectable} from "inversify";
 
-export interface BuildDao {
-    getById(id: number): Promise<Build> ;
+@injectable()
+export class BuildDao {
+
+    private webLinks: WebLinks;
+    private xmlParser: XmlParser;
+
+    constructor(@inject(TYPES.WebLinks) webLinks: WebLinks,
+                @inject(TYPES.XmlParser) xmlParser: XmlParser) {
+        this.webLinks = webLinks;
+        this.xmlParser = xmlParser;
+    }
+
+    public async getById(id: number): Promise<Build> {
+        const buildXml = await this.webLinks.getBuildInfo(id);
+        return await this.xmlParser.parseRestBuild(buildXml);
+    }
 }
