+++ src/dal/summarydao.ts
@@ -1,7 +1,27 @@
 "use strict";
 
+import {TYPES} from "../bll/utils/constants";
 import {Summary} from "../bll/entities/summary";
+import {XmlParser} from "../bll/utils/xmlparser";
+import {VsCodeUtils} from "../bll/utils/vscodeutils";
+import {RemoteBuildServer} from "./remotebuildserver";
+import {inject, injectable} from "inversify";
 
-export interface SummaryDao {
-    get(): Promise<Summary> ;
+@injectable()
+export class SummaryDao {
+
+    private remoteBuildServer: RemoteBuildServer;
+    private xmlParser: XmlParser;
+
+    constructor(@inject(TYPES.RemoteBuildServer) remoteBuildServer: RemoteBuildServer,
+                @inject(TYPES.XmlParser) xmlParser: XmlParser) {
+        this.remoteBuildServer = remoteBuildServer;
+        this.xmlParser = xmlParser;
+    }
+
+    public async get(): Promise<Summary> {
+        const gZippedSummary: Uint8Array[] = await this.remoteBuildServer.getGZippedSummary();
+        const summeryXmlObj: string = VsCodeUtils.gzip2Xml(gZippedSummary);
+        return this.xmlParser.parseSummary(summeryXmlObj);
+    }
 }
