+++ src/inversify.config.ts
@@ -11,17 +11,13 @@ import {CommandHolder} from "./commandholder";
 import {NotificationWatcherImpl} from "./bll/notifications/notificationwatcherimpl";
 import {NotificationWatcher} from "./bll/notifications/notificationwatcher";
 import {RemoteLogin} from "./dal/remotelogin";
-import {RemoteLoginImpl} from "./dal/remoteloginimpl";
 import {RemoteBuildServer} from "./dal/remotebuildserver";
 import {RemoteBuildServerImpl} from "./dal/remotebuildserverimpl";
 import {WebLinks} from "./dal/weblinks";
-import {WebLinksImpl} from "./dal/weblinksimpl";
 import {CustomPatchSender} from "./bll/remoterun/patchsenderimpl";
 import {PatchSender} from "./bll/remoterun/patchsender";
 import {SummaryDao} from "./dal/summarydao";
 import {BuildDao} from "./dal/builddao";
-import {BuildDaoImpl} from "./dal/builddaoimpl";
-import {SummaryDaoImpl} from "./dal/summarydaoimpl";
 import {TeamCityOutput} from "./view/teamcityoutput";
 import {Output} from "./view/output";
 import {PatchManager} from "./bll/utils/patchmanager";
@@ -56,12 +52,12 @@ myContainer.bind<CredentialsStore>(TYPES.CredentialsStore).to(InMemoryCredential
 myContainer.bind<ExtensionManager>(TYPES.ExtensionManager).to(ExtensionManager);
 myContainer.bind<CommandHolder>(TYPES.CommandHolder).to(CommandHolder);
 myContainer.bind<NotificationWatcher>(TYPES.NotificationWatcher).to(NotificationWatcherImpl);
-myContainer.bind<RemoteLogin>(TYPES.RemoteLogin).to(RemoteLoginImpl);
+myContainer.bind<RemoteLogin>(TYPES.RemoteLogin).to(RemoteLogin);
 myContainer.bind<RemoteBuildServer>(TYPES.RemoteBuildServer).to(RemoteBuildServerImpl);
-myContainer.bind<WebLinks>(TYPES.WebLinks).to(WebLinksImpl);
+myContainer.bind<WebLinks>(TYPES.WebLinks).to(WebLinks);
 myContainer.bind<PatchSender>(TYPES.PatchSender).to(CustomPatchSender);
-myContainer.bind<SummaryDao>(TYPES.SummaryDao).to(SummaryDaoImpl);
-myContainer.bind<BuildDao>(TYPES.BuildDao).to(BuildDaoImpl);
+myContainer.bind<SummaryDao>(TYPES.SummaryDao).to(SummaryDao);
+myContainer.bind<BuildDao>(TYPES.BuildDao).to(BuildDao);
 myContainer.bind<PatchManager>(TYPES.PatchManager).to(PatchManager).inSingletonScope();
 myContainer.bind<XmlParser>(TYPES.XmlParser).to(XmlParser).inSingletonScope();
 myContainer.bind<CvsProviderProxy>(TYPES.CvsProviderProxy).to(CvsProviderProxy).inSingletonScope();
